Improve repository structure and code quality
Implement user authentication and task management enhancements.

* **app.js**: Remove `getpass` and `os` imports. Add error handling and validation for user inputs in the `taskForm` event listener. Import `tasks.js`.
* **auth.js**: Implement user authentication and authorization logic in `loginUser`, `registerUser`, and `logoutUser` functions. Add error handling and validation for user inputs in the `loginForm` and `registerForm` event listeners.
* **index.html**: Update script tags to include the new `tasks.js` file. Add a section for user authentication.
* **tasks.js**: Add task-related functions from `app.js`. Implement error handling and validation for task inputs. Add a feature to save tasks to local storage.
* **README.md**: Update setup instructions to reflect the new file structure. Add instructions for setting up user authentication and authorization.
* **styles.css**: Enhance the user interface by adding more styling and responsiveness. Add styles for the new user authentication section.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Setland34/task-management-app/tree/add-cdp-toolkit?shareId=XXXX-XXXX-XXXX-XXXX).
