+++ index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="styles.css">
     <script src="app.js" defer></script>
     <script src="auth.js" defer></script>
+    <script src="tasks.js" defer></script>
 </head>
 <body>
     <header>
@@ -44,6 +45,30 @@
             <ul>
                 <!-- Task items will be dynamically added here -->
             </ul>
+        </section>
+        <section id="user-authentication">
+            <h2>User Authentication</h2>
+            <div id="login-form">
+                <h3>Login</h3>
+                <form>
+                    <label for="login-username">Username:</label>
+                    <input type="text" id="login-username" name="login-username" required>
+                    <label for="login-password">Password:</label>
+                    <input type="password" id="login-password" name="login-password" required>
+                    <button type="submit">Login</button>
+                </form>
+            </div>
+            <div id="register-form">
+                <h3>Register</h3>
+                <form>
+                    <label for="register-username">Username:</label>
+                    <input type="text" id="register-username" name="register-username" required>
+                    <label for="register-password">Password:</label>
+                    <input type="password" id="register-password" name="register-password" required>
+                    <button type="submit">Register</button>
+                </form>
+            </div>
+            <button id="logout-button">Logout</button>
         </section>
     </main>
     <footer>
